next-investment-api/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
crayons==0.4.0
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.7.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
bcrypt==4.2.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
